+++ src/components/Cocktail/CocktailList.vue
@@ -9,7 +9,7 @@
         This is a list of cocktails available in your Bar Assistant server. You can search for a specific cocktails by filtering them with the tags you added or by using a search term.
     </p>
     <ais-instant-search :search-client="searchClient" index-name="cocktails:name:asc" :routing="routing">
-        <ais-configure :hitsPerPage="100" :stalledSearchDelay="200" />
+        <ais-configure :hitsPerPage="100" :stalledSearchDelay="200" :filters="filters" />
         <div class="cocktail-list-search-container">
             <ais-search-box placeholder="Type to filter cocktails..." :class-names="{ 'ais-SearchBox-input': 'form-input', 'ais-SearchBox-reset': 'cocktail-list-search-container__reset' }" />
             <ais-sort-by :items="[
@@ -21,7 +21,7 @@
                 { value: 'cocktails:date:asc', label: 'Date modified asc.' },
                 { value: 'cocktails:date:desc', label: 'Date modified desc.' },
             ]" :class-names="{ 'ais-SortBy-select': 'form-select' }" />
-            <button type="button" class="button button--input" @click.prevent="showFilterContainer = !showFilterContainer">
+            <button type="button" class="button button--input" @click.prevent="toggleShow">
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24">
                     <path fill="none" d="M0 0h24v24H0z" />
                     <path d="M6.17 18a3.001 3.001 0 0 1 5.66 0H22v2H11.83a3.001 3.001 0 0 1-5.66 0H2v-2h4.17zm6-7a3.001 3.001 0 0 1 5.66 0H22v2h-4.17a3.001 3.001 0 0 1-5.66 0H2v-2h10.17zm-6-7a3.001 3.001 0 0 1 5.66 0H22v2H11.83a3.001 3.001 0 0 1-5.66 0H2V4h4.17z" />
@@ -32,90 +32,69 @@
             <template v-slot="{ items, createURL }">
                 <div class="cocktail-current-refinements">
                     <template v-for="item in items">
-                        <template v-if="item.label == 'id'">
-                            <div class="cocktail-current-refinements__refinement">
-                                <a href="#" :href="createURL(refinement)" @click.prevent="item.refinements.forEach(r => item.refine(r))">
-                                    Specific cocktails <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="16" height="16">
-                                        <path fill="none" d="M0 0h24v24H0z" />
-                                        <path d="M12 10.586l4.95-4.95 1.414 1.414-4.95 4.95 4.95 4.95-1.414 1.414-4.95-4.95-4.95 4.95-1.414-1.414 4.95-4.95-4.95-4.95L7.05 5.636z" />
-                                    </svg>
-                                </a>
-                            </div>
-                        </template>
-                        <template v-else>
-                            <div class="cocktail-current-refinements__refinement" v-for="refinement in item.refinements">
-                                <a href="#" :href="createURL(refinement)" @click.prevent="item.refine(refinement)">
-                                    {{ handleRefinementTag(refinement) }} <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="16" height="16">
-                                        <path fill="none" d="M0 0h24v24H0z" />
-                                        <path d="M12 10.586l4.95-4.95 1.414 1.414-4.95 4.95 4.95 4.95-1.414 1.414-4.95-4.95-4.95 4.95-1.414-1.414 4.95-4.95-4.95-4.95L7.05 5.636z" />
-                                    </svg>
-                                </a>
-                            </div>
-                        </template>
+                        <div class="cocktail-current-refinements__refinement" v-for="refinement in item.refinements">
+                            <a href="#" :href="createURL(refinement)" @click.prevent="item.refine(refinement)">
+                                {{ handleRefinementTag(refinement) }} <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="16" height="16">
+                                    <path fill="none" d="M0 0h24v24H0z" />
+                                    <path d="M12 10.586l4.95-4.95 1.414 1.414-4.95 4.95 4.95 4.95-1.414 1.414-4.95-4.95-4.95 4.95-1.414-1.414 4.95-4.95-4.95-4.95L7.05 5.636z" />
+                                </svg>
+                            </a>
+                        </div>
                     </template>
-                </div>
-            </template>
-        </ais-current-refinements>
-        <ais-panel>
-            <template v-slot:default="{ hasRefinements }">
-                <div class="cocktail-list-filter-panel" style="margin-bottom: 10px;" v-show="showFilterContainer">
-                    <h4>Filter by tags:</h4>
-                    <ais-refinement-list attribute="tags" :sort-by="['name:asc']" :limit="30" operator="and">
-                        <template v-slot:item="{ item, refine, createURL }">
-                            <a :href="createURL(item.value)" class="tag tag--link" :class="{ 'tag--is-selected': item.isRefined }" @click.prevent="refine(item.value)">
-                                {{ item.label }} ({{ item.count }})
-                            </a>
-                        </template>
-                    </ais-refinement-list>
-                    <h4>Filter by glass type:</h4>
-                    <ais-refinement-list attribute="glass" :sort-by="['name:asc']" :limit="30" operator="or">
-                        <template v-slot:item="{ item, refine, createURL }">
-                            <a :href="createURL(item.value)" class="tag tag--link" :class="{ 'tag--is-selected': item.isRefined }" @click.prevent="refine(item.value)">
-                                {{ item.label }} ({{ item.count }})
-                            </a>
-                        </template>
-                    </ais-refinement-list>
-                    <h4>Rating</h4>
-                    <ais-rating-menu attribute="average_rating">
-                        <template v-slot="{ items, refine, createURL }">
-                            <a v-for="item in items" :key="item.value" :href="createURL(item.value)" class="tag tag--link" :class="{ 'tag--is-selected': item.isRefined }" @click.prevent="refine(item.value)">
-                                <span v-for="(full, index) in item.stars" :key="index">
-                                    {{ full ? '★' : '☆' }}
-                                </span>
-                                & up ({{ item.count }})
-                            </a>
-                        </template>
-                    </ais-rating-menu>
-                    <h4>Cocktail filters:</h4>
-                    <div class="cocktail-list-filter-panel__toggle-refinements">
-                        <ais-toggle-refinement attribute="user_id" :on="userId">
-                            <template v-slot="{ value, refine, createURL }">
-                                <a :href="createURL(value)" class="tag tag--link" :class="{ 'tag--is-selected': value.isRefined }" @click.prevent="refine(value)">
-                                    My cocktails
-                                    ({{ value.count || 0 }})
-                                </a>
-                            </template>
-                        </ais-toggle-refinement>
-                        <ais-toggle-refinement attribute="id" :on="favoritedCocktailsIds">
-                            <template v-slot="{ value, refine, createURL }">
-                                <a :href="createURL(value)" class="tag tag--link" :class="{ 'tag--is-selected': value.isRefined }" @click.prevent="refine(value)">
-                                    My favorites
-                                    ({{ value.count || 0 }})
-                                </a>
-                            </template>
-                        </ais-toggle-refinement>
-                        <ais-toggle-refinement attribute="id" :on="shelfCocktailIds">
-                            <template v-slot="{ value, refine, createURL }">
-                                <a :href="createURL(value)" class="tag tag--link" :class="{ 'tag--is-selected': value.isRefined }" @click.prevent="refine(value)">
-                                    Shelf cocktails
-                                    ({{ value.count || 0 }})
-                                </a>
-                            </template>
-                        </ais-toggle-refinement>
+                    <div class="cocktail-current-refinements__refinement" v-for="filter in activeFilters">
+                        <a href="#" @click.prevent="filter.isActive = false">
+                            {{ filter.label }} <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="16" height="16">
+                                <path fill="none" d="M0 0h24v24H0z" />
+                                <path d="M12 10.586l4.95-4.95 1.414 1.414-4.95 4.95 4.95 4.95-1.414 1.414-4.95-4.95-4.95 4.95-1.414-1.414 4.95-4.95-4.95-4.95L7.05 5.636z" />
+                            </svg>
+                        </a>
                     </div>
                 </div>
             </template>
-        </ais-panel>
+        </ais-current-refinements>
+        <div class="cocktail-list-filter-panel hidden">
+            <h4>Cocktail filters:</h4>
+            <div class="cocktail-list-filter-panel__toggle-refinements">
+                <ais-toggle-refinement attribute="user_id" :on="userId">
+                    <template v-slot="{ value, refine, createURL }">
+                        <a :href="createURL(value)" class="tag tag--link" :class="{ 'tag--is-selected': value.isRefined }" @click.prevent="refine(value)">
+                            My cocktails
+                            ({{ value.count || 0 }})
+                        </a>
+                    </template>
+                </ais-toggle-refinement>
+                <a v-for="(customFilter, index) in filtersConfig" href="#" class="tag tag--link" :class="{ 'tag--is-selected': customFilter.isActive }" @click.prevent="toggleArrayFiltersConfig(index)">
+                    {{ customFilter.label }} ({{ customFilter.values.length }})
+                </a>
+            </div>
+            <h4>Filter by tags:</h4>
+            <ais-refinement-list attribute="tags" :sort-by="['name:asc']" :limit="30" operator="and">
+                <template v-slot:item="{ item, refine, createURL }">
+                    <a :href="createURL(item.value)" class="tag tag--link" :class="{ 'tag--is-selected': item.isRefined }" @click.prevent="refine(item.value)">
+                        {{ item.label }} ({{ item.count }})
+                    </a>
+                </template>
+            </ais-refinement-list>
+            <h4>Filter by glass type:</h4>
+            <ais-refinement-list attribute="glass" :sort-by="['name:asc']" :limit="30" operator="or">
+                <template v-slot:item="{ item, refine, createURL }">
+                    <a :href="createURL(item.value)" class="tag tag--link" :class="{ 'tag--is-selected': item.isRefined }" @click.prevent="refine(item.value)">
+                        {{ item.label }} ({{ item.count }})
+                    </a>
+                </template>
+            </ais-refinement-list>
+            <h4>Rating</h4>
+            <ais-rating-menu attribute="average_rating">
+                <template v-slot="{ items, refine, createURL }">
+                    <a v-for="item in items" :key="item.value" :href="createURL(item.value)" class="tag tag--link" :class="{ 'tag--is-selected': item.isRefined }" @click.prevent="refine(item.value)">
+                        <span v-for="(full, index) in item.stars" :key="index">
+                            {{ full? '★': '☆' }}
+                        </span>
+                        & up ({{ item.count }})
+                    </a>
+                </template>
+            </ais-rating-menu>
+        </div>
         <ais-infinite-hits>
             <template v-slot="{ items, refineNext, isLastPage }">
                 <CocktailGridContainer v-slot="observer">
@@ -134,7 +113,6 @@
 <script>
 import { instantMeiliSearch } from '@meilisearch/instant-meilisearch';
 import { history as historyRouter } from 'instantsearch.js/es/lib/routers';
-import { singleIndex as singleIndexMapping } from 'instantsearch.js/es/lib/stateMappings';
 import Auth from '@/Auth.js';
 import ApiRequests from '@/ApiRequests.js'
 import CocktailGridItem from './CocktailGridItem.vue'
@@ -153,12 +131,51 @@ export default {
             ),
             routing: {
                 router: historyRouter(),
-                stateMapping: singleIndexMapping('cocktails:name:asc'),
+                stateMapping: {
+                    stateToRoute(uiState) {
+                        const indexUiState = uiState['cocktails:name:asc'];
+
+                        return {
+                            q: indexUiState.query,
+                            tags: indexUiState.refinementList && indexUiState.refinementList.tags,
+                            glass: indexUiState.refinementList && indexUiState.refinementList.glass,
+                            my_cocktails: indexUiState.toggle && indexUiState.toggle.user_id != null,
+                            avg_rating: indexUiState.ratingMenu && indexUiState.ratingMenu.average_rating
+                        }
+                    },
+                    routeToState(routeState) {
+                        return {
+                            ['cocktails:name:asc']: {
+                                query: routeState.q,
+                                refinementList: {
+                                    tags: routeState.tags,
+                                    glass: routeState.glass,
+                                },
+                                ratingMenu: {
+                                    average_rating: routeState.avg_rating
+                                },
+                                toggle: {
+                                    user_id: routeState.my_cocktails
+                                }
+                            }
+                        }
+                    }
+                }
             },
-            userCocktails: [],
-            shelfCocktailIds: [],
             userId: Auth.getUser().id,
             showFilterContainer: false,
+            filtersConfig: {
+                shelf: {
+                    isActive: false,
+                    label: "Cocktails I can make",
+                    values: []
+                },
+                favorites: {
+                    isActive: false,
+                    label: "My favorites",
+                    values: []
+                },
+            },
         };
     },
     components: {
@@ -168,29 +185,48 @@ export default {
     },
     created() {
         document.title = `Cocktails \u22C5 Salt Rim`
-        const userId = Auth.getUser().id
-        ApiRequests.fetchUserCocktails(userId).then(data => {
-            this.userCocktails = data
-        })
 
-        ApiRequests.fetchShelfCocktails(true).then(data => {
-            this.shelfCocktailIds = data;
-        });
+        if (this.$route.query.favorites) {
+            this.toggleArrayFiltersConfig('favorites');
+        }
+
+        if (this.$route.query.shelf) {
+            this.toggleArrayFiltersConfig('shelf');
+        }
+
+        this.setupFilters()
     },
     computed: {
-        userCocktailIds() {
-            return this.userCocktails.map(c => c.id)
-        },
-        favoritedCocktailsIds() {
-            const ids = Auth.getUser().favorite_cocktails;
-            if (ids.length === 0) {
-                return null;
+        filters() {
+            const activeFilters = this.activeFilters;
+
+            if (activeFilters.length == 0) {
+                return '';
             }
 
-            return ids;
+            const values = activeFilters.map(filter => filter.values).reduce((arr1, arr2) => [...arr1, ...arr2])
+
+            return `id IN [${values.join(', ')}]`;
+        },
+        activeFilters() {
+            return Object.values(this.filtersConfig).filter(c => c.isActive);
         }
     },
     methods: {
+        toggleShow() {
+            // Instantsearch has some weird issues with refinements if using v-show
+            document.querySelector('.cocktail-list-filter-panel').classList.toggle('hidden')
+        },
+        setupFilters() {
+            ApiRequests.fetchShelfCocktails(true).then(data => {
+                this.filtersConfig.shelf.values = data;
+            });
+
+            this.filtersConfig.favorites.values = Auth.getUser().favorite_cocktails;
+        },
+        toggleArrayFiltersConfig(key) {
+            this.filtersConfig[key].isActive = !this.filtersConfig[key].isActive
+        },
         handleRefinementTag(ref) {
             if (ref.attribute == 'user_id') {
                 return `My cocktails`;
@@ -207,8 +243,9 @@ export default {
     padding: 20px;
     background-color: rgba(255, 255, 255, .5);
     margin-top: 10px;
+    margin-bottom: 10px;
     border-radius: 10px;
-    box-shadow: 0 3px 0 var(--color-bg-dark);
+    box-shadow: 0 3px 0 var(--clr-red-300);
 }
 
 .cocktail-list-filter-panel h4 {
@@ -231,6 +268,7 @@ export default {
     display: flex;
     flex-wrap: wrap;
     gap: 8px;
+    margin-bottom: 10px;
 }
 
 .cocktail-list-search-container {
@@ -279,7 +317,7 @@ export default {
     display: flex;
     justify-content: center;
     align-items: center;
-    background: var(--color-text);
+    background: var(--clr-gray-800);
     color: #fff;
     padding: 2px 10px;
     font-size: 0.7rem;
@@ -294,7 +332,7 @@ export default {
 .cocktail-current-refinements__refinement a:hover,
 .cocktail-current-refinements__refinement a:active,
 .cocktail-current-refinements__refinement a:focus {
-    background: var(--color-link-hover);
+    background: var(--clr-red-800);
 }
 
 .cocktail-list-filter-panel .ais-RatingMenu {
@@ -302,5 +340,10 @@ export default {
     flex-wrap: wrap;
     gap: 8px;
     margin-bottom: 15px;
+}
+
+.hidden {
+    visibility: hidden;
+    display: none;
 }
 </style>