+++ src/components/PageHeader.vue
@@ -12,14 +12,14 @@
 <style scoped>
 .site-page-header {
     display: flex;
-    border-bottom: 3px solid var(--color-bg-dark);
+    border-bottom: 3px solid var(--clr-red-300);
     padding-bottom: 20px;
     margin-bottom: 20px;
     align-items: center;
 }
 
 .site-page-header__title h2 {
-    font-family: var(--font-accent);
+    /* font-family: var(--font-accent); */
     font-weight: bold;
 }