+++ src/components/Settings/Navigation.vue
@@ -5,6 +5,7 @@
             <RouterLink :to="{name: 'settings.users'}">Users</RouterLink>
             <RouterLink :to="{name: 'settings.categories'}">Ingredient categories</RouterLink>
             <RouterLink :to="{name: 'settings.glasses'}">Glass types</RouterLink>
+            <RouterLink :to="{name: 'settings.tags'}">Tags</RouterLink>
         </template>
     </nav>
 </template>
@@ -34,7 +35,7 @@ export default {
 }
 
 .settings-nav a.router-link-exact-active {
-    background-color: var(--color-text);
+    background-color: var(--clr-gray-800);
     color: #fff;
 }
 </style>