+++ src/views/SettingsTagsView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import TagsList from '../components/Settings/TagsList.vue'
+</script>
+
+<template>
+  <main>
+    <TagsList />
+  </main>
+</template>