+++ src/views/ShelfCocktailsView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import Cocktails from '../components/Shelf/Cocktails.vue'
+</script>
+
+<template>
+  <main>
+    <Cocktails />
+  </main>
+</template>