+++ src/views/ShelfShoppingListView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import ShoppingList from '../components/Shelf/ShoppingList.vue'
+</script>
+
+<template>
+  <main>
+    <ShoppingList />
+  </main>
+</template>