+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# v1.3.0
+## New
+- Updated ingredients page
+    - Now includes filters and searching similar to cocktails
+    - You can now add ingredients to shopping list directly from the grid
+    - Added filters: Category, Origin and Strength
+- Added tag management to settings
+- Added shopping list page
+    - View all items on your shopping list
+    - You can now print your shopping list
+
+## Changes
+- Updated shelf page
+    - Now includes application stats
+    - Removed shelf cocktails, favorites and shopping list grid view
+    - Added 3 columns with: Latest cocktails, Latests shopping list items and Favorite cocktails
+- Refactored behavior of cocktail shelf and favorites filters
+- Small styling changes
+    - Updated default fonts
+    - Expanded default container width
+- Updated instantsearch routing handling
+
+## Work In Progress
+- Started work on PWA features
+    - Still needs a lot of testing
+
 # v1.2.0
 ## New
 - Added cocktail recipe printing